feat(utils): refactor stub-gen to stubs, dynamically create stubber module for import
